update python scripts, fix compatibility Python3/2

+++ python/mules.py
@@ -256,7 +256,7 @@ class MulesClient():
 
         n_columns = len(self.params['data format'])
         n_bytes = len(package)
-        n_samples = (n_bytes/size_element) / n_columns
+        n_samples = int((n_bytes/size_element) / n_columns)
         ####mesData = np.uint8(mesData) # Convert from binary to integers (not necessary pyton)
         if self.python2:
             bytes_per_element = np.flipud(np.reshape(list(bytearray(package)), [size_element,-1],order='F'))
@@ -270,7 +270,7 @@ class MulesClient():
         # Unrolls the matrix bytes_per_element, in "package_correct_order"
         # that has a size (n_bytes,1)
 
-        data_format_tags = self.params['data format']*int(n_samples)
+        data_format_tags = self.params['data format']*n_samples
         # Tags used to map the elements into their corresponding representation
         package_correct_order_char = "".join(map(chr,package_correct_order))
